+++ engines/bagel/boflib/file.cpp
@@ -156,7 +156,7 @@ ErrorCode CBofFile::Read(void *pDestBuf, int32 lBytes) {
 			pBuf = (byte *)pDestBuf;
 
 			while (lBytes > 0) {
-				nLength = (int)MIN(lBytes, CHUNK_SIZE);
+				nLength = (int)MIN(lBytes, (int32)CHUNK_SIZE);
 				lBytes -= CHUNK_SIZE;
 
 				if ((int)rs->read(pBuf, nLength) != nLength) {
@@ -188,7 +188,7 @@ ErrorCode CBofFile::Write(const void *pSrcBuf, int32 lBytes) {
 			pBuf = (const byte *)pSrcBuf;
 
 			while (lBytes > 0) {
-				nLength = (int)MIN(lBytes, CHUNK_SIZE);
+				nLength = (int)MIN(lBytes, (int32)CHUNK_SIZE);
 				lBytes -= CHUNK_SIZE;
 
 				if ((int)ws->write(pBuf, nLength) != nLength) {
